Changes ingredient#amount to be a float
This just allows for more granularity in the recipes

+++ lib/bread/recipes/ingredient.ex
@@ -4,7 +4,7 @@ defmodule Bread.Recipes.Ingredient do
   alias Bread.Recipes.Recipe
 
   schema "ingredients" do
-    field :amount, :integer
+    field :amount, :float
     field :name, :string
     belongs_to :recipe, Recipe
